pyproject.toml
[project]
name = "autonomous-pokerogue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.7.9",
    "imagehash>=4.3.2",
    "paddleocr[all]>=3.2.0",
    "paddlepaddle>=3.2.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.2",
]

[dependency-groups]
dev = [
    "ipython>=9.1.0",
    "pytest>=8.3.5",
    "ruff>=0.11.4",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
